refactor(drawer): improve readability of Drawer component and its children
