+++ apps/client/src/components/layout/Drawer/Drawer.styled.tsx
@@ -1,5 +1,42 @@
 import styled, { keyframes } from 'styled-components';
-import { Flex } from '@radix-ui/themes';
+import * as Dialog from '@radix-ui/react-dialog';
+import { Flex, Box } from '@radix-ui/themes';
+
+export const Drawer = styled(Box)``;
+
+const fadeIn = keyframes`
+  from {
+    opacity: 0;
+  }
+
+  to {
+    opacity: 1;
+  }
+`;
+
+const fadeOut = keyframes`
+  from {
+    opacity: 1;
+  }
+
+  to {
+    opacity: 0;
+  }
+`;
+
+export const DrawerOverlay = styled(Dialog.Overlay)`
+  position: fixed;
+  inset: 0;
+  background-color: rgb(0 0 0 / 75%);
+
+  &[data-state='open'] {
+    animation: ${fadeIn} 150ms ease-out;
+  }
+
+  &[data-state='closed'] {
+    animation: ${fadeOut} 150ms ease-in;
+  }
+`;
 
 const slideIn = keyframes`
   from {
@@ -21,7 +58,7 @@ const slideOut = keyframes`
   }
 `;
 
-export const Drawer = styled(Flex)`
+export const DrawerContent = styled(Flex)`
   position: fixed;
   top: 0;
   right: 0;
