+++ apps/client/src/components/layout/Drawer/DrawerBody/DrawerBody.styled.tsx
@@ -1,4 +1,4 @@
 import styled from 'styled-components';
 import { Flex } from '@radix-ui/themes';
 
-export const DrawerContent = styled(Flex)``;
+export const DrawerBody = styled(Flex)``;
