--- apps/client/src/components/layout/Drawer/DrawerOverlay/DrawerOverlay.tsx
@@ -1,7 +1,0 @@
-import * as S from './DrawerOverlay.styled';
-
-const DrawerOverlay = () => {
-  return <S.DrawerOverlay className="OVERLAY" />;
-};
-
-export default DrawerOverlay;
