--- apps/client/src/components/layout/Header/ToggleDrawerButton/ToggleDrawerButton.styled.tsx
@@ -1,4 +1,0 @@
-import styled from 'styled-components';
-import { Box } from '@radix-ui/themes';
-
-export const ToggleDrawerButton = styled(Box)``;
